# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/accounts/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "get account from user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserAccountAggregateResponse"
                        }
                    }
                }
            }
        },
        "/api/auth-pin": {
            "post": {
                "description": "Authenticates a user and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user via PIN",
                "parameters": [
                    {
                        "description": "User name and PIN",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/banner/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banner"
                ],
                "summary": "get banner from user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Banner"
                            }
                        }
                    }
                }
            }
        },
        "/api/debit-card/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Debit Card"
                ],
                "summary": "get banner from user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserDebitCardAggregateResponse"
                        }
                    }
                }
            }
        },
        "/api/set-pin": {
            "post": {
                "description": "Allows a user to set a PIN for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Set a PIN for a user",
                "parameters": [
                    {
                        "description": "User name and PIN",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: PIN set successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/greeting/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "get greeting from user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserGreeting"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.AccountBalanceResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "domain.AccountDetailResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "is_main_account": {
                    "type": "boolean"
                },
                "progress": {
                    "type": "integer"
                }
            }
        },
        "domain.AccountFlagResponse": {
            "type": "object",
            "properties": {
                "flag_id": {
                    "type": "integer"
                },
                "flag_type": {
                    "type": "string"
                },
                "flag_value": {
                    "type": "string"
                }
            }
        },
        "domain.AccountResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "dummy_col_3": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "domain.AccountWithDetailsResponse": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/domain.AccountResponse"
                },
                "balance": {
                    "$ref": "#/definitions/domain.AccountBalanceResponse"
                },
                "detail": {
                    "$ref": "#/definitions/domain.AccountDetailResponse"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AccountFlagResponse"
                    }
                }
            }
        },
        "domain.Banner": {
            "type": "object",
            "properties": {
                "banner_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dummy_col_11": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.DebitCardWithDetailsResponse": {
            "type": "object",
            "properties": {
                "border_color": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.UserAccountAggregateResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AccountWithDetailsResponse"
                    }
                }
            }
        },
        "domain.UserDebitCardAggregateResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DebitCardWithDetailsResponse"
                    }
                }
            }
        },
        "domain.UserGreeting": {
            "type": "object",
            "properties": {
                "dummy_col_2": {
                    "type": "string"
                },
                "greeting": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handler.PinRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                }
            }
        }
    }
}